Selectively expose new Types.* for 1.5.1.
Because 1.5.1 is to be a minor release, there is an argument
for not (yet) changing what ResultSetMetaData and friends will return
from getColumnType() for a timetz or timestamptz column, so leave those
to return Types.TIME or Types.TIMESTAMP as they always have. Client code
that cares can use getColumnTypeName() to distinguish time or timestamp
from timetz/timestamptz.

At the same time, there is no need to feign ignorance of
Types.TIME_WITH_TIMEZONE or TIMESTAMP_WITH_TIMEZONE if a client
asks for them, so do add them in the list recognized in Oid.c.