Fix old timestamp coercion for !integer_datetimes.
Testing the new java.time conversions under !integer_datetimes
revealed that they work, but the old existing ones weren't quite
right ... the same issues as in issue #155, but that fix wasn't
tested with !integer_datetimes, so they were overlooked for that case.

In passing, simplify the new _LocalDateTime_coerceDatum. Breaking
the multiplication by 1e6 into a six-bit exponent adjust and
separate 14-bit multiplier was unnecessarily cautious. There are
~ 20 bits we care about to the right of the radix point. All the rest
are available after the whole seconds are subtracted out, and even
multiplying by 1e9 won't need more than 21 of them (9 go straight to
the exponent). To the best of my knowledge, nobody ever made a
double precision format with a significand too narrow for 41 bits.