Have StAXResultAdapter implement XMLStreamWriter.
As the TODO comment said: it seems there is a practical difference
between the XMLEventWriter and XMLStreamWriter APIs, as the former
cannot distinguish between self-closed empty elements and the
start-and-end tag form, while the latter can. This should probably
be replaced by a class that implements XMLStreamWriter; that's just
more tedious and longwinded, with so many more methods to implement.

The interesting stuff is more or less done, so no more putting off
the tedious and longwinded.