Let timestamp values roundtrip even when infinite.
The Java classes representing them do not have a notion of
'infinity' or '-infinity' as PostgreSQL does. For now, just
produce instances of the desired classes with the naturally
converted values, which will appear as particular, crazy late
or crazy early, finite timestamps, and just make sure those
specific values convert back in the other direction to PostgreSQL
'infinity' and '-infinity'. Java code could use a query to
generate one of each object and save them, to recognize or use
infinite values later in real data.

The above works as advertised with integer_datetimes (which has
been the default for many PG versions, and is now not even
selectable). With the older, floating-point representation, it
doesn't, and given the combination of obsolescence of the float
representation and rarity of infinite values in practice, is arguably
not worth much effort to improve.

Add some tests with infinite values to the JDBC42_21 example, included
only when integer_datetimes is on.

In passing, update copyright in some files that were touched earlier
without doing that.