Make installing the unsandboxed example opt-in.
The prepareXMLTransform example function relies on Java's
TransformerFactory.newTemplates() method, and that seems to
fail unless unsandboxed, at least in newer JREs where the
"XSLTC" transformer compiler is used.

To avoid having the examples jar install an unsandboxed function
unasked, require an extra implementor tag to be set in
pljava.implementors if the function is wanted.