Share PL/Java and application classes in OpenJ9.
The OpenJ9 JVM has a facility for dynamically saving AOT-compiled
classes in a shared cache, which is even available for classes of
PL/Java application code, provided PL/Java's class loader does a
bit of coordination with J9's class sharing API.

The token generated for the J9 SharedClassTokenHelper is, for now,
fixed, which will be bad news if a jar is updated. It should be
computed in a way that ensures it is different after an update.