Derive a cache token from the entryId.
This is dead simple, and workable because (in the current, 1.5.0
sqlj schema), the entryId is a SERIAL column, and all entries are
deleted/reinserted by replace_jar. Of course this means that every
cached class from the jar is invalidated even if few have changed,
but optimizing that can be future work.