Simplify by using JSR 310 date/time classes.
They output and parse the same character format as the XML Schema types,
and the local time/timestamp classes don't have the counterintuitive
dependency on the session time zone.

Also, the decision to introduce SQLXML as a non-compatibility-breaking
change means it also has to be requested explicitly from the driver,
just like the JSR 310 types.