Include Saxon example, but not built by default.
Add a Maven profile -Psaxon-examples that will pull in Saxon-HE
as a dependency and build the Saxon example, which (unlike the
rest of the PL/Java sources) depends on Java 8.

In passing, update the version of Saxon-HE in the dependency.

For now, to avoid javadoc conniptions when the Saxon dependency
isn't present, exclude the saxon example from javadoc unconditionally.
Maybe there is a way to tie that also to the profile activation, later.