Document side-effect of 5433cd5.
Checking the shared cache before querying for the jar_entry row
saves some time, but also changes a behavior after a jar has been
replaced. replace_jar clears the class loaders and function cache
in the session where it is executed, but not in others; the historic
(and still current) behavior is that they continue executing the old
classes they have loaded, but will hit a ClassNotFoundException the
moment another class needs to be loaded from the old jar.

It is now possible they would keep right on executing the old code
successfully, if the needed class is found in the shared cache.

While that's a behavior change, it doesn't seem important to revert.
There's nothing very documented or innately desirable about the old
behavior anyway, and use of J9 sharing is opt-in. So just document it.

In passing, also moderate the recommendation of -Xquickstart; it goes
without saying that a sufficiently compute-heavy workload could see a
net slowdown because of the less optimized code, but what's striking
is how little effort that takes. So replace the recommendation with a
more carefully hedged one.