Send reader to tuning page for AppCDS setup details.
The details of getting Hotspot to share Saxon classes are
too fiddly to devote the space to in general docs, but are
covered in all their fiddly glory in a methodology section
on the performance-tuning wiki page, so direct the reader
there to learn all about them.

A bit of other minor copyediting.