Add advice about uppercasing parameter names.
The toUppercase() in SPIResultSetMetaData.getColumnLabel() seems to be
required by nothing in JDBC that I can find. I see nothing like it in
pgjdbc. But it has been in PL/Java since 2005. It can't be just yanked
out abruptly, but it's got to be deprecated and removed in some future
release.

Therefore, advise not to rely on the behavior for parameter names
passed into XQuery here, but to really, for the time being, spell them
in all upper case, both in the XQuery text and in the SQL. In PostgreSQL,
of course, that also means they have to be quoted in the SQL.