Avoid trying to use type names in Type.c messages.
The simplest (or simplest-looking) way was to call format_type_be,
but that function throws an error if the passed Oid is InvalidOid
or an unknown type, so instead of the intended useful message in
the log, a less helpful one gets thrown (and possibly turns a
debug message or warning into an error).

There is no equally-easy-to-call backend function to format a
regtype with a benign fallback if it isn't known. The fmgr-callable
format_type() works, but then you're going cstring::text::cstring
just to make a message that might be at debug level and go nowhere.

Just put the numeric oids in the message.