Preserve Java 6 buildability.
The AbstractResultSet needs to declare abstract the getObject-
with-class method that won't really appear in the interface
until Java 7. The PassXML example should use getSQLXML(), which
existed in Java 6, rather than getObject(..., SQLXML.class),
which did not.

The TypeRoundTripper example has inherent reason to use the
getObject(..., Class<?>) method, so just move it into an
example/jdk7 directory that is only built when a JDK [1.7,)
profile is active, much like what was done with the Saxon
example. But in this case, do not exclude it from the javadoc
build; I would rather say "Java 6 is too old to build the docs"
than leave more stuff out of the docs so Java 6 can build them.

Why two different older versions of maven-resources-plugin
were specified in two different places may just have been an
artifact of history. 3.0.1 is its latest version with a class
file version not greater than 50.

The JDBC42_21 example is a strange case; it tests some JDBC4.2
functionality, so it definitely won't work before Java 8, but is written
so it can still be compiled in Java 6. However, the SQL generator will
object to its 'requires' dependency on TypeRoundTripper, which got moved
to the jdk7 package, so this must go there too.