Sop for certain Rhino-based JavaScript engines.
The last assignment in this JavaScript snippet sets computedPath
to a concatenation (with +) of two strings, and Rhino handles that
in a clever way (borrowed from V8, it seems), producing an instance
of its ConsString class.
https://mozilla.github.io/rhino/javadoc/org/mozilla/javascript/ConsString.html

... which works very nicely if the JSR-223 adapter layer between
Rhino and javax.script API knows about the ConsString object and
converts transparently to a Java String. It does, in the Oracle
Java 6 and 7 builds. The adapter layer is a separate project, not
part of Rhino proper.

What apparently happened with the OpenJDK 6 and 7 is that they bundled
perfectly good versions of Rhino with somewhat incomplete versions
of the adapter, and that adapter does not know how to make a Java
String from a ConsString.

Using the ConsString to construct a new, ordinary JavaScript string
seems to solve the problem; the adapter knows what to do with one
of those.