Use script task to quote libjvmdefault as C string
This is needed because the current practice of passing the default JVM
path as a command-line macro definition to the C compiler and relying
on C preprocessor stringification to make a string literal of it only
works when the path includes no word-like components that collide with
macros in existence during the build. (Issue #176)

It seemed tidier to use scriptdef, but that plain doesn't work
from a maven-antrun-plugin configuration; the plugin writes out a
script file containing every part of the scriptdef except the script.

The bug report hasn't been assigned yet...
https://issues.apache.org/jira/browse/MANTRUN-207

too bad, as the only problem looks to be right here:
http://svn.apache.org/viewvc/maven/plugins/tags/maven-antrun-plugin-1.8/src/main/java/org/apache/maven/plugin/antrun/AntrunXmlPlexusConfigurationWriter.java?revision=1646987&view=markup#l69

So, for now, this is done with a script tag, no scriptdef.