Add a ddr.reproducible option to SQL generator.
When true, it constrains the order of the generated DDR to be
consistent over successive compilations of the same sources,
useful in distribution builds where reproducibility is valued.
When false, it selects the previous behavior, where descriptors
can have different order from one compilation to the next of the
same sources; during testing, this can help to detect whether
there are necessary provides/requires relationships that haven't
been declared.

Addresses issue #181.

The 'reproducible' option also fixes the order of Java-to-SQL
type mappings known to the generator, though this is an internal
detail; the ordered mappings are not emitted explicitly into the
DDR, and if a change in their order would affect what is written
there, that would probably indicate a more serious issue.