Have Loader force SPI read_only false as needed.
This is meant to eliminate the hacks needed for classloading initiated
from deployment descriptors in the course of install_jar or replace_jar,
as detailed in issue #178.

It is not the only possible approach, but was favored in
https://www.postgresql.org/message-id/14247.1537418670%40sss.pgh.pa.us

An approach that could be less invasive and tidier was proposed in
https://www.postgresql.org/message-id/5BA6BA96.5030309%40anastigmatix.net
but would presuppose later PostgreSQL API such as PushCopiedSnapshot,
which does not appear until 9.1. On this branch, PL/Java still supports
earlier versions, so this approach will have to do.