Avoid timestamp in SetOfRecordTest.
The SetOfRecordTest is only there to demonstrate that sets of RECORD
can be returned, and that casts are done from a ResultSetHandle as
needed. Using a timestamp as one of the test values can make it also
(in some circumstances) a demonstration of java.sql.Timestamp's
unintuitive dependence on local time zone settings ... which is why
the JDBC 4.2 mapping to java.time.LocalDateTime is preferable, which
the documentation and release notes already explain. The purpose of
SetOfRecordTest can be achieved just as well using types that don't
have that sort of environment dependency.