Use schema class loader in Oid.getJavaClass
Add a Function_getCurrentLoader() to retrieve a saved reference
to the initiating loader used to find the function's containing
class. That should be the "schema loader" for the schema in which
the function is declared. Tempting to cheat and just use
getClassLoader on the function's class, but that would return the
defining loader, in cases where it and the initiating loader are
different.

The choice to use the schema loader for the innermost executing
PL/Java function is for consistency with the documented convention
(in the developer notes) of using the innermost executing function's
schema to determine the type map to use.

Taking a step back, the whole Oid class seems like something between
a misnomer and a solution-in-search-of-a-problem. It is something of
a stash for information on the correspondences between PG and Java
types, but hardly a complete one, and a lot of that information is in
other places; meanwhile, plenty of things in PG, besides types, have
Oids. In some future version, I envision this Oid class gone, with its
type mapping information merged into something more like a much-evolved
TypeBridge that would centralize a lot of type-correspondence knowledge
that is held now in many places. There might be an entirely different
class to represent oids, if that is useful, with subclasses
corresponding to the different reg* oid types in PG.

But for now, the existing Oid class gets nursed along.