Add a test for issue #192.
It's sufficient to use the TypeRoundTripper to ask for the JDBC class
name for a column of type point (after the Point example has installed
itself as a mirror of that type). The call to ResultSetMetaData's
getColumnClassName() fails, because it uses Oid.getJavaClass, which
tries to use the class loader that loaded Oid itself, not a schema
loader that could see the installed jar providing Point.

In passing, remove some unnecessary alias names from column definition
lists in the documentation and tests. At some point, I must have been
convinced they were needed, but this code is intended as an example, so
shouldn't be needlessly cluttered.