(Belatedly) advance version to 1.5.2-SNAPSHOT.
This should have been part of the preceding commit, as it just
followed a full release of 1.5.1.

I do not seem to be holding to the plan announced in 2353235, to
keep the version string unchanged all through a RELx_y_STABLE branch.
(The plan only lasted until f087234). It would indeed simplify merges
from the branch back to master. But it hardly makes sense for a version
string to stay x.y.z-SNAPSHOT after x.y.z has been released.

In the next RELx_y_STABLE branch, perhaps the convention should be
that the SNAPSHOT version string is simply x.y-SNAPSHOT. That could
then be left alone through the life of the branch.