Repair java.sql.Date breakage from 72cbb15.
That commit was part of the work of adding the java.time types for 1.5.1
(issue #137). A potential for overflow in timezone handling for timestamps
was corrected by changing the API of Timestamp_getTimeZone_id(). There
was known to be one other use of that function, in Date.c. A comment was
even added in Timestamp.h advising that there was a known use in Date.c.
But somehow updating that known use in Date.c slipped through the cracks.

The end result was that the new, preferable java.time API got added
(preferable in part because it eliminates the unexpected timezone
dependencies when manipulating local dates and times), but in the process
a spooky timezone-dependent off-by-a-day error got introduced to the
old date <-> java.sql.Date conversions.