Handle the retirement of magical oids.
Upstream 578b229 turns the oid columns of system catalogs into
ordinary columns, to be retrieved in the ordinary way by name or
column number. Non-system-catalogs don't get to have system-managed
oids at all any more. (You could give a table an oid-typed column
named oid and make a key of it, but no function is provided to
automatically assign a value to such a column.)