Add an SQL_ASCII encoding.
This is a principled Java take on the PostgreSQL definition of
SQL_ASCII as an encoding where the seven-bit ASCII values are
themselves and the remaining eight-bit values are who-knows-what.
It isn't appropriate to just copy byte values with no conversion,
as that would amount to saying we know the values correspond to
LATIN-1, which would be lying. Java strings are by definition Unicode,
so it's not ok to go stuffing code points in that do not mean what
Unicode defines those code points to mean.

What this decoder does is decode the seven-bit ASCII values as
themselves, and decode each eight-bit value into a pair of Unicode
noncharacters, one from the range u+fdd8 to u+fddf, followed by one
from u+fde0 to u+fdef, where the first one's four low bits are the
four high bits of the original value, and the second has the low four.
The encoder transparently reverses that.

Those noncharacter code points are permanently defined in Unicode
to have no glyphs, no correspondence to specific characters, and
no interesting properties. Implementing this charset allows PL/Java
code to work usefully in a database with SQL_ASCII encoding, when the
expectation is that whatever the code needs to recognize, act on, or
edit will be in ASCII, and any non-ASCII content can be passed along
uninterpreted and unchanged.