The allow-java-encodings feature doesn't work.
The property is recognized, and is set in the property manager.
But com.sun.org.apache.xerces.internal.impl.XMLEntityManager has
overloaded reset methods, one taking a PropertyManager, and one
taking a ComponentManager. Only the second one honors the feature.
XMLStreamReaderImpl.reset() calls the first one, natch.

But all is not lost, as createXMLStreamReader with an encoding
parameter seems to accept it without caring about its IANA-or-Java-ness.
It seems a bit redundant to both arrange to have the proper encoding
declared in the stream and to pass it to createXMLStreamReader, but
it gets the job done.