SQLXML nits when server encoding != UTF-8
StAX will object to reading a stream that declares a non-IANA-registered
encoding--even if Java has a corresponding Charset--unless an obscure
feature allow-java-encodings has been set. Also, the StAXResultAdapter
needs to flush the XMLStreamWriter before closing the underlying stream;
writing in UTF-8 encoding did not expose that, but writing in another
encoding can.

When writing the PostgreSQL XML data type, throw conformance to the wind
and omit any declaration of the encoding in use, even when not UTF-8.
Some of the PostgreSQL core functions (IS DOCUMENT for sure) can
misbehave if the named encoding isn't something libxml approves of, even
when it is the correct name of the server encoding in use.

Improve the PassXML example to show that, when writing a SQLXML object
and not through the SAX, StAX, or DOM API, it is the application's
responsibility to arrange for the stream to be in the server encoding.