Prepare to migrate to DualState from other idioms.
Older code where a Java object is associated with some native state has
been using some combination of Object.finalize() (deprecated for removal
in newer Java environments), JavaWrapper, Invocation_createLocalWrapper,
with a lot of details scattered in a lot of places, and not a lot of
explanation.

The DualState mechanism introduced with VarlenaWrapper should be able
to subsume all of those cases, eliminate the use of Object.finalize,
and reduce (maybe eliminate?) the use of JNI from threads other than
the main PG thread.

Prepare the way by declaring that the "resource owner" associated with
a DualState does not have to be only a PG ResourceOwner. It can also be
the address of some other sort of allocated struct that has a scope-like
behavior. In PG 9.5 and later, a MemoryContext with a
MemoryContextCallback could serve, and PL/Java's currentInvocation can
be used to scope a state to the currently-executing PL/Java function.
Unless the memory allocator's broken, such alternative allocated
addresses won't be mistakable for actual ResourceOwners, so no special
magic should be needed for DualState objects scoped in any of those ways
to coexist.

In passing, change the queue implementation for s_liveInstances in
DualState. LinkedBlockingDeque was more heavyweight than needed; it
was chosen in haste after starting with ConcurrentLinkedDeque and then
finding it wasn't in Java 6, but ConcurrentLinkedQueue is, and has all
the capabilities needed. Its iterator().remove() method really runs in
constant amortized time, as it just leaves the unlinking to be done on
a subsequent traversal, which happens regularly in our case.

If it can be proven that no manipulation of s_liveInstances will happen
except on one thread interacting with PG, it could be overkill to be
using a concurrent queue at all.

Get rid of HeapTupleHeader.java, which has been dead since a4f6c9e.