HeapTupleHeader -> SingleRowReader
The architecture has been confusing for a while around HeapTupleHeader
(with a long-abandoned Java class and still-used .h/.c files simply
providing functions used for SingleRowReader and SQLInputFromTuple,
but with some of the associated native functions found in Composite.c).

Collect all that stuff into SingleRowReader.[hc], to back the
SingleRowReader Java class in the typical way. SQLInputFromTuple
can simply be a subclass of that, implementing SQLInput.