DualState-ify Relation.
Relation was handled with Invocation_createLocalWrapper, that is,
as something with a lifetime scoped to the current PL/Java function
invocation. That's been duplicated, using DualState with
currentInvocation passed as the resource owner. But it could be worth
investigating how and where Relation objects are actually getting
constructed, to see if that's the right scope for all cases.