An optimization for non-transient native scopes.
DualState works when NULL is passed as a resource owner, in which
case the native state had better be in some long-lived area like
JavaMemoryContext, because no call to nativeRelease with a real pointer
will ever match it, so javaStateUnreachable or javaStateReleased will
be the only ways it can be freed. (It follows that any DualState flavor
that will be used that way had better have a freeing action as part
of its javaStateUnreachable / javaStateReleased.)

It seems preferable to keep things in more tightly scoped contexts and
give them real owners, but this pattern was used in enough of the
former-JavaWrapper cases just migrated that it's worth supporting
(at least until further analysis can migrate some of those to use
appropriate scopes).

So, a worthwhile optimization is to track objects that lack resource
owners in a simpler data structure, that does not need to support any
retrieval by resource owner. They still have to be kept live somehow
(otherwise Java forgets about them and will not enqueue them when their
referents are unreachable), but a simple IdentityHashMap suffices, and
they can be removed in O(1) when found unreachable or released by Java.

For objects with resource owners, the tracking structure is now a hash
map from owner to doubly-linked list of the associated instances. The
list is directly implemented with prev and next references in the
instances themselves, rather than using a Collections class, so that
when an instance is found unreachable or released from Java, it also can
be removed from the list in O(1) with no searching. Exit of the native
resource owner scope is, of course, handled by looking up the owner in
the map and processing every instance remaining on the list.

The structures used now are plain vanilla java.util implementations with
no claimed thread safety properties (and the same is true of the
directly-implemented doubly-linked list code). The Javadoc now explains
at length the usage requirements that must be observed for the design to
be safe.