Factor out 'specialization' check in SQL generator.
This will make it easier to support idioms where a Java
parameter or return is typed A<B> (or some C that directly or
indirectly extends A<B>) and the generator needs to find B.

Fixes a latent (in this branch) bug where parameterized return
types could be written into the AS string in all their glory,
instead of properly as their erasures. It's latent in this branch
because the only use made of parameterized return types is for
SETOF with Iterator<T>, and the return type isn't written for that
case, but would become a real bug with any future addition of other
idioms involving parameterized return types.

First done in an explorative branch against master, with this
generally-useful refactoring rebased here. A couple other minor
differences from master are also included here, to keep future
merges simpler.