Use JNI header constant field values.
Neither the javah manual page, nor the Java 8+ javac manual page,
actually comes out and says that static final primitive constants in a
class get #defines created for them in the JNI header, but they do.

The new-in-Java-8 @Native annotation for fields might suggest that
it is needed to get a field copied into the .h file, but really seems
only to ensure that a header is generated even if the class has
no native methods. As long as a header is generated, it seems to get
all the static final primitive fields (that have constant-expression
initializers) whether so annotated or not.

Therefore, take advantage of those to avoid divergent values between
Java and C, either by directly using the #defines from Java in the C
code, or by using StaticAssertStmt to confirm the values match the
existing C constants they are intended to match.

In passing, update the SPI return codes to include the newer ones.