DualState-ify Portal
As far as I can discern, the complexity in Portal.c involving a
hash map from native to Java Portal objects was only there to support
hooking the Portal cleanup callback for detecting native release.
That callback is supplied the native Portal pointer, so the map was
needed for the callback to find the corresponding Java object. I could
find no indication that it served as a cache, or that a Portal_create
call ever returned a retrieved object rather than a new one.

As it happens, a Portal object has a ResourceOwner, so it should fit
rather directly into the DualState scheme, with no need for such
complexity of its own.

The resource owner callback happens a bit later (at drop of the Portal,
where the cleanup callback could happen as soon as it reached DONE or
FAILED state). A portal can't be executed again after DONE or FAILED
and will protest with an error message if that's attempted, so using
the resource owner may leave a gap in which Java still has the Portal
reference but the Portal will protest with appropriate errors logged
if Java tries to use it. But a non-dangling pointer whose use leads
to appropriate errors being logged seems tolerable, and we will know
about the Portal's release in time to invalidate a dangling pointer.