New approach to concurrency for DualState.
To date, Java synchronized blocks have been made to serve the purpose
of concurrency management for DualState objects, but the fit has been
awkward at best. If the concurrency needs for DualState are viewed
on a clean slate:

- a DualState mates up some state in PostgreSQL (with lifetime/validity
  determined by PostgreSQL mechanisms like resource owners and memory
  contexts) with some Java state (with lifetime/validity determined by
  Java mechanisms).

- Java can have many threads, but PostgreSQL only supports one (or one
  at a time, depending on the pljava.java_thread_pg_entry setting).

- JNI calls that will enter PostgreSQL routines must be serialized onto
  one thread (or one at a time), but direct examination of a chunk of
  native memory, or even JNI calls to thread-safe, non-PostgreSQL
  routines, ought to be permitted from other threads, as long as the
  native state being examined is expected not to change during the
  access.

- That requires mechanisms for not just data visibility across threads,
  but for exclusion. A thread must be able to 'pin' some native state
  briefly while peeking at it, and any code that handles a PostgreSQL
  callback that is about to change or release native state must use a
  locking operation that will block for any existing pin.

That leads to a structure like a vanilla reader/writer lock, where the
read lock (hereafter called a 'pin') can be taken by any number of
threads, while the write lock (hereafter simply 'lock') can only be held
by one thread, is blocked by any existing pins, and blocks any later
requested pins. For this application, a further simplification is that
it makes sense to limit all operations that could need the lock to run
on one particular thread (the one for which Backend.threadMayEnterPG()
would return true), so there is no need to support multiple conflicting
requests for locks.

Every DualState instance needs to remember when its native state has
been invalidated, or its Java state has been explicitly freed or found
unreachable. These conditions can be checked atomically as part of any
pin request, so pin() will throw the appropriate exceptions if the state
should not be accessed, saving client code some extra checks and TOCTTOU
races. If the pin succeeds, access to the state is ok while it is held.

The Java synchronization mechanisms offered in java.util.concurrent do
not include anything with exactly these behaviors, but do include the
lower-level primitives on which it can be built, as is done here.