Simplify PgSavepoint.
This does refer to a bit of what could be called native state,
but not really enough to bother adapting it to DualState. Just
simplify it instead. The PostgreSQL functions it relies on only
need two integers (transaction ID and nest level), so those can
be kept right in the Java object.

It's very tempting to eliminate the special-case code in
Invocation.java for remembering PgSavepoints, and the special code
in SPIConnection to call the special code in Invocation to remember
and forget PgSavepoints, and simply migrate PgSavepoint to using
DualState like most other stuff, and have the Invocation as its
resource owner. But because of the way savepoints nest, it is
sufficient to do what the special-case code now does, and remember
only the outermost savepoint that is set, so the effort of migrating
to a different mechanism doesn't offer much benefit.

It appears that at one time Thomas factored out the savepoint
set/release/rollback logic into functions and a struct that could
be offered upstream to core as new SPI_ functions, but that was a
long time ago and it looks as if core didn't take up the offer. So,
just merge the logic back into PgSavepoint so it isn't scattered all
over creation. There are still the mildly incestuous arrangements
involving SubXactListener, Invocation, and SPIConnection mentioned
above.

Both JDBC and PostgreSQL's BeginInternalSubTransaction allow for a
savepoint to be unnamed, so no longer treat the no-arg setSavepoint
as if it were setSavepoint("anonymous_savepoint").