Guarantee threadMayEnterPG for subxact callbacks.
The behavior of PgSavepoint.forId() presupposes that the callbacks
are made on the PG thread, which is perfectly reasonable to suppose,
but could be exposed to surprises in java_thread_pg_entry=allow mode.
This change will make the callbacks behave as if in =block mode.

Perhaps that is too drastic a change for 1.5.x, and the C code
should just make two upcalls, one to forId() and then to the actual
listener, with only the first one locked?