Simplify SubXactListener the same way.
It is a tiny bit more involved, because the ids of savepoints
are passed. To resolve the ids to PgSavepoint objects, the C code
will upcall to the Java forId() method with the Backend.THREADLOCK
monitor held (as in java_thread_pg_entry = block mode), so that
method can get away with an assert Backend.threadMayEnterPG().

The upcall that actually calls the listeners will be made with the
monitor released, as that's The Way It's Been Done. This still is
the PG thread, but without the monitor held, Backend.threadMayEnterPG()
won't know it, so the Java code shouldn't blindly assert it.