Adopt generics in more places where applicable.
A solid, if non-exhaustive, effort to bring the code up to date with
Java generics and transparent boxing/unboxing, based on a simplistic
mechanical search for places where unnecessary explicit casting/boxing/
unboxing was being done.

In the files touched, also elide a bunch of 'this.' qualifications on
instance method calls. And fix the spelling of 'Coercion' in three methods.

Bonus: fix three very old thinkos in SPIPreparedStatement that no one
ever seems to have run into or reported.