Build on Java versions with and without javah
Java 8 added a -h option to javac to allow generation of .h files
from Java sources in one step, subsuming (most of) the function of
the older javah command. Java 10 duly axed javah, leaving only the
javac -h option, a stroke of brilliance that left no way to h a class
file, such as java.sql.Types, for which no source file is present.

The smart-alecky workaround would be to add a new <scriptdef> in the
build process that would iterate java.sql.Types.class.getDeclaredFields()
and write them to an .h file. It would hardly be difficult (at least
for this special case, a class defining only some int constants and
loadable at build time with no dependency issues).

But that hardly seems necessary, considering that (as of Java 12) there
have been no changes to java.sql.Types constants since Java 8, there is
only one PL/Java source file (type/Oid.c) that uses them, and that file
will probably go away anyway in a foreseeable PL/Java release. So here,
a version of java_sql_Types.h generated on Java 8 is simply provided in
include/fallback/jdbc. If it is not gone by the time new Types constants
are added in a future Java version, they can be added to it then.

There is some conditional compilation in Oid.c based on this file, so
continue to generate it when possible (using javah unless the build JDK
is 10 or newer), switching to javac -h and the fallback copy only for 10
or newer.