Allow a parameter's SQL name to be explicit.
The SQL generator has been emitting CREATE FUNCTION with parameter
names that match the Java names of the function parameters, so that
the SQL functions can be called with named-parameter notation.

But there can be cases when a different SQL name is wanted for the
parameter, whether because the Java name happens to be a reserved
SQL word, or because a particular SQL parameter name is wanted (comes
from a standard, maybe?) and happens to be a reserved Java word,
or just because.

Further overload the per-parameter @SQLType annotation (which is
already a little overloaded and misnamed, being able to supply
default values as well as type information) with a name property,
which can be a regular identifier or a delimited one (in case a
parameter name is really wanted with the same spelling as an SQL
reserved word).

Copy in the ISO_DELIMITED_IDENTIFIER regex from the master branch,
to allow checking that ... nothing else is really checked that
carefully yet, but everything has to start somewhere.