Accomodate Saxon 9.8 to 9.9 API changes
A few small changes to allow depending on the latest, 9.9 series
Saxon releases.

XQueryEvaluator.setContextItem() can now throw a checked exception
(in the case of a query that declares a context item, and does not
declare it to be external).

The change of the return type of XdmValue.iterator() from
XdmSequenceIterator to Iterator<XdmItem> seems to have not been
meant for release. The method still returns an XdmSequenceIterator
(confirmed in the source), so the best solution appears to be a
narrowing cast, rather than any change to the expected return type:

https://sourceforge.net/p/saxon/mailman/message/36716685/