Correct issue #200 issues in Time.c
Almost! But does the "out-of-range values passed to
java.sql.Time constructor" turn out to be NOTABUG? Consider
passing current_time to Java, with the PG session and Java
default time zones the same, but the current status of summer
time being different from that of 1970-Jan-1. The right mSecs
value gets set and things are good, but the java.sql.Time
object won't report the same hour as PG. What a mess.

No regression test added yet, owing to the fiddliness of
changing default timezones ... can be revisited.