Now use Java time zone for timestamp
Eliminates the need for the exposed Timestamp_getTimeZone_id()
function and its hairy two-microsecond units. Can also eliminate
the bitshifted manipulations in Timestamp_coerceObjectTZ_id():
that was a thinko in 72cbb15. If mSecs could be anywhere near
overflowing, we'd be in deep trouble anyway, because it has to
end up as microseconds and fit in the same bit width.