Update user docs pertaining to date/time
The bug/notabug debate in issue #200 and the commit notes for 4a0bb84
is hereby resolved by executive decision: bug. The old approach
technically violated the contract of java.sql.Time, and differed from
pgJDBC's behavior in the same case.

A back-compatibility mode would not be infeasible if requested, but
would be a bit of work to do well (the former behavior could avoidably
fail on days with time zone rule transitions), too much work to invest
in advance of a request for it, on a deprecated mapping where clients
really ought to migrate to java.time.