Method for timezone masquerading
The thing that needs to happen when a PostgreSQL value with no time zone
is mapped to one of the java.sql.{Date,Time,Timestamp} classes (which
all implicitly have a time zone) could be described as 'masquerading'
the zoneless PostgreSQL value as a UTC value that will be rendered by
Java as similarly as possible to the original. The 'unmasking' operation
needed in the reverse direction is similar but simpler; both can be
combined in one method.

The code formerly doing that appeared in a few places without very
clearly conveying what was going on, and was using the wrong time zone
to do it (the PostgreSQL session time zone when it needs to be the Java
runtime's default time zone, see issue #200, though of course the two
are often the same).

Replace that code with a single new method utcMasquerade with more
explicit comments explaining the need for, constraints on, and
limitations of the 'masquerade'.